+++ src/main/java/com/braintreegateway/Dispute.java
@@ -42,10 +42,32 @@ public class Dispute {
         UNRECOGNIZED;
     }
 
+    // NEXT_MAJOR_VERSION Remove this enum
+    /**
+     * @deprecated use ProtectionLevel enum instead
+     */
+    @Deprecated
     public enum ChargebackProtectionLevel {
         EFFORTLESS,
         STANDARD,
         NOT_PROTECTED;
+    }
+
+    public enum ProtectionLevel {
+        EFFORTLESS_CBP("Effortless Chargeback Protection tool"),
+        STANDARD_CBP("Chargeback Protection tool"),
+        NO_PROTECTION("No Protection");
+
+        private final String value;
+
+        ProtectionLevel(String value) {
+            this.value = value;
+        }
+
+        @Override
+        public String toString() {
+            return value;
+        }
     }
 
     private final Calendar createdAt;
@@ -70,7 +92,9 @@ public class Dispute {
     private final Reason reason;
     private final Status status;
     private final Kind kind;
-    private final ChargebackProtectionLevel chargebackProtectionLevel;
+    @Deprecated
+    private final ChargebackProtectionLevel chargebackProtectionLevel; // Deprecated
+    private final ProtectionLevel protectionLevel;
     private final BigDecimal amount;
     private final BigDecimal disputedAmount;
     private final BigDecimal wonAmount;
@@ -96,6 +120,17 @@ public class Dispute {
         status = EnumUtils.findByName(Status.class, node.findString("status"), Status.UNRECOGNIZED);
         kind = EnumUtils.findByName(Kind.class, node.findString("kind"), Kind.UNRECOGNIZED);
         chargebackProtectionLevel = EnumUtils.findByName(ChargebackProtectionLevel.class, node.findString("chargeback-protection-level"), ChargebackProtectionLevel.NOT_PROTECTED);
+        switch(String.valueOf(node.findString("chargeback-protection-level"))) {
+            case "effortless":
+                protectionLevel = ProtectionLevel.EFFORTLESS_CBP;
+                break;
+            case "standard":
+                protectionLevel = ProtectionLevel.STANDARD_CBP;
+                break;
+            default:
+                protectionLevel = ProtectionLevel.NO_PROTECTION;
+        }
+
         amount = node.findBigDecimal("amount");
         disputedAmount = node.findBigDecimal("amount-disputed");
         wonAmount = node.findBigDecimal("amount-won");
@@ -196,8 +231,16 @@ public class Dispute {
         return kind;
     }
 
+    /**
+     * @deprecated use getProtectionLevel() instead
+     */
+    @Deprecated
     public ChargebackProtectionLevel getChargebackProtectionLevel() {
         return chargebackProtectionLevel;
+    }
+
+    public ProtectionLevel getProtectionLevel() {
+        return protectionLevel;
     }
 
     public BigDecimal getAmount() {