+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## Unreleased
+## 3.19.0
+* Deprecate `chargebackProtectionLevel` and add `protectionLevel` to `Dispute` and `DisputeSearchRequest`
+* Add `networkTokenizationAttributes` to `TransactionCreditCardRequest`
+* Add `NetworkTokenizationAttributesRequest` class
+* Add validation error code `CREDIT_CARD_NETWORK_TOKENIZATION_ATTRIBUTE_CRYPTOGRAM_IS_REQUIRED`
+* Add 'updateCustomFields' to `Transaction`
 
 ## 3.18.0
 * Add `ExternalVaultCardRequest` class